Extract profile utilities
