+++ genres_modul.py
@@ -1,5 +1,6 @@
 # Version 0.1.8
 import os, json, shutil
+from profiles_utils import data_path, load_profiles
 from PyQt5.QtWidgets import (
     QApplication, QWidget, QVBoxLayout, QHBoxLayout, QLabel,
     QLineEdit, QPushButton, QListWidget, QMessageBox, QComboBox,
@@ -7,18 +8,6 @@ from PyQt5.QtWidgets import (
 )
 from PyQt5.QtGui import QIcon
 from PyQt5.QtCore import Qt
-
-def data_path():
-    return os.path.join(os.path.dirname(__file__), "Projekt", "genres_profile.json")
-
-def load_profiles():
-    path = data_path()
-    if not os.path.exists(path):
-        os.makedirs(os.path.dirname(path), exist_ok=True)
-        with open(path, "w", encoding="utf-8") as f:
-            json.dump({"Favoriten":[]}, f)
-    with open(path, "r", encoding="utf-8") as f:
-        return json.load(f) or {"Favoriten":[]}
 
 def save_profiles(d):
     with open(data_path(), "w", encoding="utf-8") as f:
