+++ zufallsgenerator_modul.py
@@ -1,22 +1,11 @@
 # Version 0.1.8
-import random, os, json
+import random
+from profiles_utils import data_path, load_profiles
 from PyQt5.QtWidgets import (
     QWidget, QVBoxLayout, QHBoxLayout, QLabel,
     QPushButton, QComboBox, QListWidget, QMessageBox, QApplication
 )
 from PyQt5.QtCore import Qt
-
-def data_path():
-    return os.path.join(os.path.dirname(__file__), "Projekt", "genres_profile.json")
-
-def load_profiles():
-    path = data_path()
-    if not os.path.exists(path):
-        os.makedirs(os.path.dirname(path), exist_ok=True)
-        with open(path, "w", encoding="utf-8") as f:
-            json.dump({"Favoriten":[]}, f)
-    with open(path, "r", encoding="utf-8") as f:
-        return json.load(f) or {"Favoriten":[]}
 
 class ZufallsGeneratorModul(QWidget):
     def __init__(self):
